Add port speed to physical network plugin
The port_speed_physnet plugin provides a mechanism for assigning a
physical network to ironic ports based on the speed of the interface
reported by the extra-hardware collector.

The plugin is configured via the option [port_physnet]
port_speed_mapping, which is a comma-separated list of <port
speed>:<physical network> tuples.