Add Link Aggregate physical network plugin
The lag_physnet plugin populates the physical_network field of ironic
ports that are members of a Link Aggregate Group (LAG), as determined by
LLDP data gathered by the discovery ramdisk.

The plugin is configured via the option [port_physnet] lag_physnet,
which is the name of the physical network to apply.