Fix AttributeError when some ports don't have physnet
The base_physnet plugin base class checks the port list in the node
cache to see if a physical_network field is set on the first port. It
appears that in some cases the field might be set on the first port but
not on one or more other ports. This change fixes this by checking all
ports.

Fixes: #7